alias: add diskus , update setup.sh tmux: update soruce

+++ alias/setup.sh
@@ -1,12 +1,26 @@
 #!/usr/bin/env sh
-BASEDIR=$(cd "$(dirname "$1")" && pwd)/$(basename "$1")
+BASEDIR=$(cd "$(dirname "$0")" && pwd)
 
-#add .aliases to .bashrc
-ln -sf $BASEDIR/.aliases ~/.aliases
+# Create symbolic links 
+if [ ! -f ~/.aliases ]; then
+    ln -sf $BASEDIR/.aliases ~/.aliases
+fi
 
+# Check .zshrc & .bashrc
+CONFIG_FILES="$HOME/.bashrc $HOME/.zshrc"
+for config in $CONFIG_FILES; do
+    #avoid empty file
+    if [ -f $config ]; then
+        # Install these to $config
+        grep "if \[ -f ~\/.aliases ]; then" $config >/dev/null
+        if [ $? -ne 0 ]; then
+        cat <<-EOF >> $config
+
+# source at ~/.dotfiles/alias/.aliases
 if [ -f ~/.aliases ]; then
     . ~/.aliases
 fi
-
-#add .aliases to .zshrc
-source $HOME/.aliases
+EOF
+        fi
+    fi
+done
